+++ modules/http4s/test/src-ce3/compliancetests/package.scala
@@ -1,0 +1,63 @@
+/*
+ *  Copyright 2021-2022 Disney Streaming
+ *
+ *  Licensed under the Tomorrow Open Source Technology License, Version 1.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *     https://disneystreaming.github.io/TOST-1.0.txt
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import cats.effect.unsafe.IORuntime
+import cats.effect.IO
+import smithy4s.dynamic.DynamicSchemaIndex
+import smithy4s.Schema
+import smithy4s.Document
+import smithy4s.http.PayloadError
+import smithy4s.http.CodecAPI
+import smithy4s.dynamic.model.Model
+import smithy4s.dynamic.DynamicSchemaIndex.load
+import smithy4s.schema.Schema.document
+package object compliancetests {
+
+  private def loadDynamic(
+      doc: Document
+  ): Either[PayloadError, DynamicSchemaIndex] = {
+    Document.decode[Model](doc).map(load)
+  }
+
+  private def decodeDocument(
+      bytes: Array[Byte],
+      codecApi: CodecAPI
+  ): Document = {
+    val schema: Schema[Document] = document
+    val codec: codecApi.Codec[Document] = codecApi.compileCodec(schema)
+    codecApi
+      .decodeFromByteArray[Document](codec, bytes)
+      .getOrElse(sys.error("unable to decode smithy model into document"))
+  }
+
+  private lazy val path = sys.env
+    .get("MODEL_DUMP")
+    .map(fs2.io.file.Path(_))
+    .getOrElse(sys.error("MODEL_DUMP env var not set"))
+
+  private implicit val runtime: IORuntime = cats.effect.unsafe.IORuntime.global
+  def dynamicSchemaIndexLoader(codecAPI: CodecAPI): DynamicSchemaIndex =
+    fs2.io.file
+      .Files[IO]
+      .readAll(path)
+      .compile
+      .toVector
+      .map(_.toArray)
+      .map(decodeDocument(_, codecAPI))
+      .map(loadDynamic(_).getOrElse(sys.error("unable to load Dynamic path")))
+      .unsafeRunSync()
+
+}
